Rename RelatedCollections to EntityGroup, RelatedSection to RelatedCollectionsCard, EntityRelatedCollections to EntityRelatedCollectionsCard
