+++ packages/portal/src/components/entity/EntityRelatedCollectionsCard.vue
@@ -3,11 +3,11 @@
     v-if="relatedCollections.length > 0 || entityUris.length > 0"
     class="text-left related-collections-card"
   >
-    <RelatedCollections
+    <EntityGroup
       :related-collections="relatedCollections"
       :entity-uris="entityUris"
       data-qa="related collections"
-      @fetched="handleRelatedCollectionsFetched"
+      @fetched="handleRelatedCollectionsCardFetched"
     />
   </b-card>
 </template>
@@ -16,13 +16,13 @@
   import { BASE_URL as EUROPEANA_DATA_URL } from '@/plugins/europeana/data';
   import { getEntityUri, getEntityQuery, getEntityTypeApi, normalizeEntityId } from '@/plugins/europeana/entity';
 
-  import RelatedCollections from '../related/RelatedCollections';
+  import EntityGroup from '@/components/entity/EntityGroup';
 
   export default {
-    name: 'EntityRelatedCollections',
+    name: 'EntityRelatedCollectionsCard',
 
     components: {
-      RelatedCollections
+      EntityGroup
     },
 
     props: {
@@ -100,7 +100,7 @@
     },
 
     methods: {
-      handleRelatedCollectionsFetched(relatedCollections) {
+      handleRelatedCollectionsCardFetched(relatedCollections) {
         this.relatedCollections = relatedCollections;
         this.$emit('fetched', relatedCollections);
       }
