+++ packages/portal/src/pages/exhibitions/_exhibition/index.vue
@@ -72,7 +72,7 @@
             cols="12"
             class="col-lg-8"
           >
-            <RelatedCollections
+            <EntityGroup
               :entity-uris="relatedLink"
             />
           </b-col>
@@ -100,7 +100,7 @@
       AuthoredHead: () => import('../../../components/authored/AuthoredHead'),
       ContentWarningModal: () => import('@/components/generic/ContentWarningModal'),
       RelatedCategoryTags: () => import('@/components/related/RelatedCategoryTags'),
-      RelatedCollections: () => import('@/components/related/RelatedCollections')
+      EntityGroup: () => import('@/components/entity/EntityGroup')
     },
     mixins: [
       exhibitionChapters,
